Rename image lib biome provider to v2
