+++ common/addons/biome-provider-image-v2/src/main/java/com/dfsek/terra/addons/biome/image/v2/config/converter/ClosestBiomeColorConverterTemplate.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.addons.biome.image.config.converter;
+package com.dfsek.terra.addons.biome.image.v2.config.converter;
 
 import com.dfsek.tectonic.api.config.template.annotations.Value;
