+++ common/addons/biome-provider-image-v2/src/main/resources/terra.addon.yml
@@ -4,7 +4,7 @@ contributors:
 id: biome-provider-image
 version: @VERSION@
 entrypoints:
-  - "com.dfsek.terra.addons.biome.image.ImageBiomeProviderAddon"
+  - "com.dfsek.terra.addons.biome.image.v2.ImageBiomeProviderAddon"
 website:
   issues: https://github.com/PolyhedralDev/Terra/issues
   source: https://github.com/PolyhedralDev/Terra
